完成资源管理功能
Former-commit-id: 3b8af0327f3a7e6992a071230ec617d73a1d7f85
Former-commit-id: f510d842350917ab38b579e3e0951c898c399a98

+++ include/BDXKEditor/Editor/Core/GUITransferer.h
@@ -38,7 +38,7 @@ namespace BDXKEditor
         std::string fieldPath;
         std::string fieldName;
         bool isInPath = false;
-        size_t isNested = -1;
+        int isNested = -1;
         std::function<void(const ObjectPtrBase&)> clickObjectEvent;
 
         CustomTransferFunc(int, TransferInt)
